+++ R/fPCpretreatment.R
@@ -13,7 +13,7 @@ filter_seasons <- function(las, months = 1:12, gpstime_ref = "2011-09-14 01:46:4
   proportions_of_winter_pont <- (1 - nrow(las@data) / length(months_acquisition)) * 100
   if (proportions_of_winter_pont > 0) {
     if (plot_hist_days) {
-      hist(
+      graphics::hist(
         datetime,
         breaks = "day",
         main = "Histogram of acquisition date", xlab = "Date of acquisition"
@@ -39,7 +39,7 @@ filter_date_mode <- function(las, deviation_days = Inf, gpstime_ref = "2011-09-1
     return(las)
   }
 
-  hist_test <- hist(
+  hist_test <- graphics::hist(
     datetime,
     breaks = "day",
     plot = plot_hist_days,
@@ -75,56 +75,13 @@ filter_date_mode <- function(las, deviation_days = Inf, gpstime_ref = "2011-09-1
   return(las)
 }
 
-add_traj_from_las <- function(las) {
-  .N <- X <- Y <- Z <- gpstime <- ReturnNumber <- NULL
-  las_4_traj <- las
-  traj <- try(
-    lidR::track_sensor(las_4_traj, algorithm = lidR::Roussel2020()),
-    silent = TRUE
-  )
-  if (nrow(lidR::filter_ground(las)) == 0) {
-    warning("Only ground points in the tile. NULL returned")
-    return(NULL)
-  }
-  if (class(traj)[1] == "try-error") {
-    first_last <- lidR::filter_firstlast(las_4_traj)
-    tab_count <- first_last@data[, .(count = .N), by = gpstime]
 
-    las_4_traj@data <- las_4_traj@data[!gpstime %in% tab_count[count > 2]$gpstime]
-
-    traj <- try(lidR::track_sensor(las_4_traj, thin_pulse_with_time = 0, algorithm = lidR::Roussel2020()), silent = T)
-  }
-  # if track sensor not working at all take mean coordinates ( 1400 for Z) and gpstime to estimate trajectory
-  if (class(traj)[1] == "try-error") {
-    traj <- data.table::data.table(lidR::filter_ground(las)@data[, 1:4])
-    traj <- traj[, .(Easting = mean(X), Northing = mean(Y), Elevation = mean(Z) + 1400, Time = mean(gpstime)), ]
-  }
-  if (class(traj)[1] != "data.table") {
-    traj <- data.table::data.table(cbind(sf::st_coordinates(traj), Time = traj$gpstime))
-  }
-  # if track sensor not working  at all take mean coordinates ( 1400 for Z) and gpstime to estimate trajectory
-
-  if (nrow(traj) == 0) {
-    traj <- data.table::data.table(lidR::filter_ground(las)@data[, 1:4])
-    traj <- traj[, .(Easting = mean(X), Northing = mean(Y), Elevation = mean(Z) + 1400, Time = mean(gpstime)), ]
-  }
-  names(traj) <- c("Easting", "Northing", "Elevation", "Time")
-  # Find closest gpstime between traj and las
-  nn2_gpstimes <- RANN::nn2(traj$Time, las@data$gpstime, k = 1)
-  las@data <- cbind(las@data, traj[nn2_gpstimes$nn.idx, ])
-
-  las <- lidR::add_lasattribute(las, name = "Easting", desc = "traj")
-  las <- lidR::add_lasattribute(las, name = "Northing", desc = "traj")
-  las <- lidR::add_lasattribute(las, name = "Elevation", desc = "traj")
-  las <- lidR::add_lasattribute(las, name = "Time", desc = "aeroplane time")
-
-  las
-}
 
 #' Point cloud pre-treatment for using fCBDprofile_fuelmetrics in pixels
 #'
 #' @description Function for preprocessing las (laz) files for use in fCBDprofile_fuelmetrics. This can be used in the catalog_apply lidR function. The pretreatment consists of normalizing the point cloud and adding various attributes: Plane position for each point (easting, northing, elevation), LMA (leaf mass area) and wood density (WD) by intersecting the point cloud with an LMA and WD map or by providing LMA and WD values.
-#' @param chunk character. path to a las (laz) file. Can be apply to a catalog see lidr catalog apply) # nolint: line_length_linter.
+#' @param chunk LAS object, LAScluster chunk or character. If character, the path to a las (laz) file is expected.
+#' Can be apply to a catalog see lidR::catalog_apply # nolint: line_length_linter.
 #' @param classify logical (default is FALSE). Make a ground classification. Only if the original point cloud is not classified
 #' @param LMA character or numeric. Default = 140. If available path to a LMA map (.tif) of the area if available or a single LMA value in g.m² (e.g 140. cf: Martin-Ducup et al. 2024)
 #' @param WD character or numeric. Default = 591. If available, path to a WD map (.tif) of the area if available or a single WD value in kg.m3 (e.g 591 cf: Martin-Ducup et al. 2024).
@@ -138,9 +95,15 @@ add_traj_from_las <- function(las) {
 #' It is expected to be in timezone UTC. Default is "2011-09-14 01:46:40" which is the standard GPS Time (1980-01-06 00:00:00)
 #' plus 1e9 seconds, as defined in LAS 1.4 specifications.
 #' @param season_filter numeric. A vector of integer for months to keep (e.g: 5:10 keep retunrs between may and october)
-#' @return a Normalized point cloud (.laz) with several new attributes need to run fCBDprofile_fuelmetrics
+#' @param traj sf object. Trajectory covering the LAS chunk. The sf object is expected to have a column gpstime and Point Z geometries.
+#' If NULL, the function will try to compute it.
+#' @return LAS object. A normalized point cloud (.laz) with several new attributes needed to run fCBDprofile_fuelmetrics, see details.
 #' @details
-#' The attributes added to the laz are LMA : LMA value of each point. Zref :original Z; Easting, Northing, Elevation, Time that are the X,Y,Z position of the plane and the its GPStime for each point (obtained from lidR::track_sensor()). In a following version it will be possible to directly load a trajectory file if available.
+#' The attributes added to the laz are:
+#' - LMA : LMA value of each point
+#' - Zref :original Z
+#' - Easting, Northing, Elevation, Time that are the X,Y,Z coordinates of the sensor
+#'   and the its GPStime for each point (obtained from lidR::track_sensor()).
 #' @examples
 #' \donttest{
 #' path2laz <- system.file("extdata", "M30_FontBlanche.laz", package = "lidarforfuel")
@@ -149,8 +112,8 @@ add_traj_from_las <- function(las) {
 #' # displaying the new attributes in the las
 #' names(M30_FontBlanche_pretreated)
 #' }
-#' @export
 #' @import data.table
+#' @export
 fPCpretreatment <- function(
   chunk,
   classify = FALSE,
@@ -163,24 +126,36 @@ fPCpretreatment <- function(
   start_date = "2011-09-14 01:46:40",
   season_filter = 1:12,
   deviation_days = Inf,
-  plot_hist_days = FALSE
+  plot_hist_days = FALSE,
+  traj = NULL
 ) {
-
   X <- Z <- Classification <- NULL
 
   # read chunk
-  las <- lidR::readLAS(chunk)
+  if (inherits(chunk, "LAS")) {
+    las <- chunk
+  } else {
+    las <- lidR::readLAS(chunk)
+  }
   # TODO: filter virtual points, cf virtual classes in lidar-hd specs
   las <- filter_seasons(las, season_filter, plot_hist_days = plot_hist_days)
   las <- filter_date_mode(las, deviation_days, plot_hist_days = plot_hist_days)
-  # does it really happens in lidar-hd?
+  # TODO: does it really happens in lidar-hd?
   las <- lidR::filter_poi(las, !is.na(X))
 
   if (lidR::is.empty(las)) {
     return(NULL)
   }
 
-  las <- add_traj_from_las(las)
+  if (is.null(traj)) {
+    warning(paste0(
+      "Computing trajectory from LAS file...\n",
+      "Trajectory would better be computed outside pretreatment, ",
+      "with a buffer (e.g. 500m) to avoid border effects."
+    ))
+    traj <- get_traj(las, thin = 0.0001, interval = .2, rmdup = TRUE, renum = TRUE)
+  }
+  las <- add_traj_to_las(las, traj)
 
   if (classify == TRUE) {
     lidR::classify_ground(las, algorithm = lidR::csf())
@@ -201,18 +176,16 @@ fPCpretreatment <- function(
 
   # LMA
   if (is.numeric(LMA)) {
-    las@data$LMA <- LMA
-  }
-  if (is.numeric(WD)) {
-    las@data$WD <- WD
-  }
-  if (is.numeric(LMA) == FALSE) {
+    las@data[["LMA"]] <- LMA
+  } else {
     ## Load LMA map
     LMA_map <- terra::rast(LMA)
     ### Add LMA to point cloud
     las <- lidR::merge_spatial(las, LMA_map$LMA, attribute = "LMA")
   }
-  if (is.numeric(WD) == FALSE) {
+  if (is.numeric(WD)) {
+    las@data[["WD"]] <- WD
+  } else {
     ## Load LMA map
     WD_map <- terra::rast(WD)
     ### Add WD to point cloud
@@ -221,7 +194,7 @@ fPCpretreatment <- function(
 
   # Normalyze height
   las <- lidR::normalize_height(las = las, algorithm = lidR::tin())
-  # Remove points too low (<-3) or too high (>35m). Keep vegetation, soil, non classified and water point only
+  # Remove points too low (<-3) or too high (>35m). Keep vegetation, ground, unclassified and water point only
   las <- lidR::filter_poi(las, (Classification <= 5 | Classification == 9) & Z < Height_filter)
   las <- lidR::classify_noise(las, lidR::sor(5, 10))