+++ tests/testthat/test-fPCpretreatment.R
@@ -1,4 +1,4 @@
-testthat::test_that("filter_season", {
+test_that("filter_season", {
   plot_hist_days <- FALSE
   ref <- as.POSIXct("2011-09-14 01:46:40", tz = "UTC")
   gpstime <- seq.Date(as.Date("2023-07-01"), as.Date("2023-08-31"), by = "days")
@@ -12,24 +12,24 @@ testthat::test_that("filter_season", {
   # here returned las should be full
   months <- 5:10
   las1 <- filter_seasons(las, months, plot_hist_days = plot_hist_days)
-  testthat::expect_true(nrow(las1@data) == nrow(las@data))
+  expect_true(nrow(las1@data) == nrow(las@data))
 
   # returned las should be half size
   months <- 7
-  testthat::expect_warning({
+  expect_warning({
     las1 <- filter_seasons(las, months, plot_hist_days = plot_hist_days)
-    testthat::expect_true(nrow(las1@data) == nrow(las@data) / 2)
+    expect_true(nrow(las1@data) == nrow(las@data) / 2)
   })
 
   # returned las should be half empty
   months <- 1:5
-  testthat::expect_warning({
+  expect_warning({
     las1 <- filter_seasons(las, months, plot_hist_days = plot_hist_days)
   })
-  testthat::expect_true(nrow(las1@data) == 0)
+  expect_true(nrow(las1@data) == 0)
 })
 
-testthat::test_that("filter_date_mode", {
+test_that("filter_date_mode", {
   # scales well (tested for 1e7 points)
   N <- 9
   # N+1 points have the same date
@@ -41,7 +41,7 @@ testthat::test_that("filter_date_mode", {
 
   deviation_days <- 2
 
-  testthat::expect_warning({
+  expect_warning({
     las <- filter_date_mode(
       las,
       deviation_days = deviation_days,
@@ -49,7 +49,7 @@ testthat::test_that("filter_date_mode", {
       plot_hist_days = FALSE
     )
   })
-  testthat::expect_true(nrow(las@data) == deviation_days * 2 + N + 1)
+  expect_true(nrow(las@data) == deviation_days * 2 + N + 1)
 
   # test where only one date is present, should not filter any point
   data[["gpstime"]] <- runif(nrow(data)) * 3600 * 24
@@ -60,21 +60,38 @@ testthat::test_that("filter_date_mode", {
     gpstime_ref = "2023-01-01 00:00:00",
     plot_hist_days = FALSE
   )
-  testthat::expect_true(nrow(las@data) == nrow(data))
+  expect_true(nrow(las@data) == nrow(data))
 })
 
 
-testthat::test_that("fPCpretreatment", {
-  path2laz <- system.file(
-    "extdata", "M30_FontBlanche.laz",
-    package = "lidarforfuel"
-  )
+test_that("fPCpretreatment", {
+  path2laz <- small_laz_file()
   # LMA value selected = 120.6 that is the LMA for Pinus halepensis, the
   # dominant species of the plot
-  m30_font_blanche_pretreated <- fPCpretreatment(path2laz, LMA = 120.6)
+  expect_warning(
+    m30_font_blanche_pretreated <- fPCpretreatment(path2laz, LMA = 120.6)
+  )
   # displaying the new attributes in the las
   expected_names <- c(
     "Easting", "Northing", "Elevation", "Time", "LMA", "WD", "Zref"
   )
-  testthat::expect_contains(names(m30_font_blanche_pretreated), expected_names)
+  expect_contains(names(m30_font_blanche_pretreated), expected_names)
 })
+
+
+
+# test_that("fPCpretreatment-full-tile", {
+#   path2laz <- "/media/DATA/boissieu/git/LidarForFuel/draft/data/lidarhd.old/LHD_FXX_0904_6339_PTS_LAMB93_IGN69.copc.laz"
+#   # LMA value selected = 120.6 that is the LMA for Pinus halepensis, the
+#   # dominant species of the plot
+#   tic("Total time")
+#   expect_warning(
+#     m30_font_blanche_pretreated <- fPCpretreatment(path2laz, LMA = 120.6)
+#   )
+#   toc()
+#   # displaying the new attributes in the las
+#   expected_names <- c(
+#     "Easting", "Northing", "Elevation", "Time", "LMA", "WD", "Zref"
+#   )
+#   expect_contains(names(m30_font_blanche_pretreated), expected_names)
+# })